Migrate ReadGwlfDataFile module
In the VB6 version, this module is responsible for initializing
variables and populating values from a GMS file. Since the GMS file is
already loaded into a Python object in our version, some things had to
be changed.

To avoid "resetting" values that were loaded from a GMS file, most
initializations have been moved to the `defaults` function of
the `Model` class.

I did not move the remaining initializations because their array bounds
depend on instance data and I haven't figured out the best way to handle
this yet.

Connects #6