Merge model defaults with parser
The primary reason for this change is to ensure that all arrays parsed
from a GMS file use the correct data type (numpy arrays instead of
Python lists).

In the process of making this change, I decided that the parser should
return an instance of `DataModel` instead of a dictionary. The purpose of
this is mainly to reduce noise and make the code look a bit cleaner.

Also, there was a side-discussion where it was decided that gwlf-e
should not assume default values when we run the model from MMW. This
makes the default values in `DataModel` redundant. However, we still
need these default values when loading existing GMS files, so I moved
them to the parser class.

Connects #13
Connects #15
Connects #23