Fix LE values
Previously, this line of code would raise a `ZeroDivisionError` when
`StreamFlowVol[Y, i]` is 0, because of the negative exponent. I tried to
fix this by checking for 0 first, but it caused a lot of missing data.

Looking at a more up-to-date version of the GWLF-E code, I noticed they
had a much better solution. Replacing the -1.6 exponent with 0.6
accomplishes the objective and avoids a `ZeroDivisionError`.