Add support for Windows environments
Due to https://github.com/numba/numba/issues/3304 the current
setup did not work in Windows environments. That issue has a
fix in https://github.com/numba/numba/pull/3329, which is
available in numba==0.41.0, so we upgrade our dependencies to
that.

In some cases, llvmlite has to be installed manually as a
pre-requisite as well, so the README is amended to instruct so.